+++ src/libAtomVM/nifs.h
@@ -34,19 +34,19 @@
 extern "C" {
 #endif
 
-#define VALIDATE_VALUE(value, verify_function) \
-    if (UNLIKELY(!verify_function((value)))) { \
-        argv[0] = ERROR_ATOM;                  \
-        argv[1] = BADARG_ATOM;                 \
-        return term_invalid_term();            \
+#define VALIDATE_VALUE(value, verify_function)        \
+    if (UNLIKELY(!verify_function((value)))) {        \
+        context_set_exception_class(ctx, ERROR_ATOM); \
+        ctx->exception_reason = BADARG_ATOM;          \
+        return term_invalid_term();                   \
     }
 
-#define RAISE_ERROR(error_type_atom)   \
-    do {                               \
-        ctx->x[0] = ERROR_ATOM;        \
-        ctx->x[1] = (error_type_atom); \
-        return term_invalid_term();    \
-    } while (0);
+#define RAISE_ERROR(error_type_atom)                  \
+    do {                                              \
+        context_set_exception_class(ctx, ERROR_ATOM); \
+        ctx->exception_reason = (error_type_atom);    \
+        return term_invalid_term();                   \
+    } while (0)
 
 const struct Nif *nifs_get(const char *mfa);